requirements.txt
# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.12

# LangChain & LangGraph - Compatible versions
langchain==0.3.13
langchain-openai==0.2.14
langgraph==0.2.59
langchain-core==0.3.28

# OpenAI
openai==1.59.8

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1

# CLI Tools
httpx==0.27.2
rich==13.9.4